Doodle Studio 95 Intergration
The plan is to use doodle studio to give the game more charm. I want to make it so that you clean a 3d blender model, and then it changes to a doodle object form. Using change animation could help, but it really depends (if im going from blender to doodle, then i can just use an if statement like in the test scene)

next up is raycast. i would do it now but i been at work staring at this computer for like hours