Fixing Logic, Raycasts implemented
The raycasts are now working as intended. Now I do not need to use distance to trigger functions.

All that is left is to make sure that things work fine visually (progress bar) and logically (dialogue box). I figured it out with dialog box one, so it should work fine for everything else